# internal/http/router.go
package router

import (
	_ "embed"
	"net/http"

	"github.com/gin-gonic/gin"

	openapi "github.com/snsilvam/kaizensnsilvam-backend/docs"
	"github.com/snsilvam/kaizensnsilvam-backend/internal/http/handlers"
	"github.com/snsilvam/kaizensnsilvam-backend/internal/http/middleware"
)

//go:embed swagger/index.html
var swaggerUI []byte

func New(familyHandler *handlers.FamilyHandler, userHandler *handlers.UserHandler, incomeHandler *handlers.IncomeHandler, dashboardHandler *handlers.DashboardHandler, allowedOrigins []string) *gin.Engine {
	r := gin.New()

	r.Use(gin.Logger())
	r.Use(gin.Recovery())
	r.Use(middleware.CORS(allowedOrigins))

	// Sin esto Gin responde 404 al preflight de rutas con método no registrado
	// y el navegador nunca ve los headers de CORS.
	r.HandleMethodNotAllowed = true
	r.NoRoute(func(c *gin.Context) { c.Status(http.StatusNotFound) })

	r.GET("/health", handlers.Health)
	swagger := func(c *gin.Context) {
		c.Data(http.StatusOK, "text/html; charset=utf-8", swaggerUI)
	}
	r.GET("/", func(c *gin.Context) {
		c.Redirect(http.StatusTemporaryRedirect, "/swagger")
	})
	r.GET("/swagger", swagger)
	r.GET("/swagger/", swagger)
	r.GET("/docs/swagger.yaml", func(c *gin.Context) {
		c.Data(http.StatusOK, "application/yaml; charset=utf-8", openapi.SwaggerYAML)
	})

	families := r.Group("/families")
	{
		families.POST("", familyHandler.Create)
		families.GET("/:id", familyHandler.GetByID)
	}

	users := r.Group("/users")
	{
		users.POST("", userHandler.Create)
		users.GET("/:id", userHandler.GetByID)
	}

	incomes := r.Group("/incomes")
	{
		incomes.POST("", incomeHandler.Register)
	}

	r.GET("/dashboard", dashboardHandler.Get)

	return r
}